requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
python-multipart
SpeechRecognition
pyttsx3
pyaudio
